requirements.txt
openai
zhipuai
numpy
python-dotenv
torch
torchvision
torchaudio
transformers
tqdm==4.66.3
PyPDF2
markdown
html2text
tiktoken
beautifulsoup4
sentencepiece==0.1.99
modelscope
requests
json5
protobuf
einops
Requests==2.31.0